Authentication and authorization using JWT for posting and deleting posts

+++ controllers/api/v1/posts_api.js
@@ -1,4 +1,4 @@
-const { findById } = require('../../../models/post');
+// const { findById } = require('../../../models/post');
 const Posts = require('../../../models/post');
 const Comment = require('../../../models/comment');
 
@@ -24,24 +24,26 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Posts.findById(req.params.id);
 
-        // if (post.user = req.user.id) {
+        if (post.user == req.user.id) {
 
-        post.remove();
+            post.remove();
 
-        await Comment.deleteMany({ post: req.params.id });
+            await Comment.deleteMany({ post: req.params.id });
 
-        return res.json(200, {
-            message: "Post has been deleted"
-        });
+            return res.json(200, {
+                message: "Post has been deleted"
+            });
 
-        // else {
-        // return res.redirect('back');
-        // }
+        } else {
+            return res.json(401, {
+                message: "You can not delete this post!"
+            });
+        }
 
     } catch (err) {
         console.log(err);
         return res.json(500, {
             message: "Internal Server Error"
-        })
+        });
     }
 }
